refactor(ava/ckb): migrate ckb to ava v3 package
