--- packages/ckb/__tests__/index.test.ts
@@ -1,7 +1,0 @@
-import { CKBJson, CKBOptions, addChart } from '../src/index';
-
-test('knowledge', () => {
-  expect(!!CKBJson()).toBe(true);
-  expect(!!CKBOptions()).toBe(true);
-  expect(typeof addChart).toBe('function');
-});
