--- packages/ckb/src/index.ts
@@ -1,29 +1,0 @@
-export { CKBJson, addChart } from './pack';
-export { CKBOptions } from './options';
-export { Language } from './i18n';
-export { CHART_ID_OPTIONS } from './chartID';
-
-export type { TransKnowledgeProps } from './i18n/interface';
-export type { ChartID } from './chartID';
-export type {
-  ChartKnowledge,
-  DataPrerequisite,
-  LevelOfMeasurement,
-  Family,
-  Purpose,
-  CoordinateSystem,
-  GraphicCategory,
-  Shape,
-  Channel,
-  RecommendRating,
-  ChartKnowledgeJSON,
-  FAMILY_OPTIONS,
-  PURPOSE_OPTIONS,
-  COORD_TYPE_OPTIONS,
-  GRAPHIC_CATEGORY_OPTIONS,
-  SHAPE_OPTIONS,
-  LOM_OPTIONS,
-  CHANNEL_OPTIONS,
-  RECOMMEND_RATING_OPTIONS,
-  DataPrerequisiteJSON,
-} from './interface';
